# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "AuthError": {
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaylistSummary": {
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_public": {
            "default": false,
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "track_count": {
            "title": "Track Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "track_count"
        ],
        "title": "PlaylistSummary",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sf_session": {
        "description": "HttpOnly JWT session cookie set by the login or registration endpoint.",
        "in": "cookie",
        "name": "sf_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LoggeRythm API",
    "version": "1.0.0",
    "x-loggerythm-compatible-contract-versions": [
      "v1"
    ],
    "x-loggerythm-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/plays": {
      "post": {
        "operationId": "record_play_api_me_plays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Track"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Record Play",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/playlists": {
      "get": {
        "operationId": "list_playlists_api_playlists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSummary"
                  },
                  "title": "Response List Playlists Api Playlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "List Playlists",
        "tags": [
          "playlists"
        ]
      },
      "post": {
        "operationId": "create_playlist_api_playlists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Create Playlist",
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/playlists/public": {
      "get": {
        "description": "Public, non-empty playlists from other users — for the community shelf.",
        "operationId": "public_playlists_api_playlists_public_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "sf_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sf Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSummary"
                  },
                  "title": "Response Public Playlists Api Playlists Public Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Playlists",
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/playlists/{playlist_id}": {
      "delete": {
        "operationId": "delete_playlist_api_playlists__playlist_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Delete Playlist",
        "tags": [
          "playlists"
        ]
      },
      "get": {
        "operationId": "get_playlist_api_playlists__playlist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "sf_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sf Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Playlist",
        "tags": [
          "playlists"
        ]
      },
      "patch": {
        "operationId": "update_playlist_api_playlists__playlist_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Update Playlist",
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/playlists/{playlist_id}/cover": {
      "get": {
        "operationId": "get_cover_api_playlists__playlist_id__cover_get",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cover",
        "tags": [
          "playlists"
        ]
      },
      "put": {
        "operationId": "set_cover_api_playlists__playlist_id__cover_put",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_cover_api_playlists__playlist_id__cover_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Set Cover",
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/playlists/{playlist_id}/export": {
      "get": {
        "description": "Download all tracks in a visible playlist as an ordered MP3 ZIP.",
        "operationId": "export_playlist_api_playlists__playlist_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Export Playlist",
        "tags": [
          "playlists"
        ]
      }
    },
...
    "/api/search/playlist": {
      "get": {
        "operationId": "search_playlist_api_search_playlist_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSearchResult"
                  },
                  "title": "Response Search Playlist Api Search Playlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Playlist",
        "tags": [
          "browse"
        ]
      }
    },
...
  }
}